Add CreateAsync static methods and obsolete all HttpStream constructors
Doing so, we can avoid doing dangerous [sync over async][1] when accessing the [Stream.Length][2] for the first time since the `HttpStream.StreamLength` will already be known.

The code path will be

```
CacheStream.Length
  Wait
    GetLengthAsync
      GetStreamLengthOrDefault
```

The `await WrapTask(readAsync(0L, new byte[1], 0, 1, CancellationToken.None));` will **not** be called so `return func().Result` is safe to call at that point since we know that the `Result` is already computed.

[1]: https://github.com/davidfowl/AspNetCoreDiagnosticScenarios/blob/master/AsyncGuidance.md#warning-sync-over-async
[2]: https://docs.microsoft.com/en-us/dotnet/api/system.io.stream.length